[Fabric] Fix BinSkim errors in preparation for 1ES Template migration (#12295)
## Description

This PR fixes the various BinSkim errors detected by running against the configuration of BinSkim mandated by the 1ES template.

Resolves: #12286

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
To fix the issues with our code so we can make out code publishing compliant.

Resolves: #12286

### What

This PR enables Spectre Mitigation, Control Flow Guard, and CET Shadow Stack compiler/linker flags for all supported configurations, as well as removing the warning suppressions used by various Fabric-only files and making the appropriate fixes.

## Screenshots
N/A

## Testing
N/A

## Changelog
Should this change be included in the release notes: yes

Fixed BinSkim errors to be compliant with publishing standards

+++ vnext/Microsoft.ReactNative/Fabric/Composition/TextInput/WindowsTextInputState.cpp
@@ -21,10 +21,10 @@ WindowsTextInputState::WindowsTextInputState(
     AttributedString attributedString,
     AttributedString reactTreeAttributedString,
     ParagraphAttributes paragraphAttributes,
-    float defaultThemePaddingStart,
-    float defaultThemePaddingEnd,
-    float defaultThemePaddingTop,
-    float defaultThemePaddingBottom)
+    double defaultThemePaddingStart,
+    double defaultThemePaddingEnd,
+    double defaultThemePaddingTop,
+    double defaultThemePaddingBottom)
     : mostRecentEventCount(mostRecentEventCount),
       attributedString(std::move(attributedString)),
       reactTreeAttributedString(std::move(reactTreeAttributedString)),
